% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{age_standardize}
\alias{age_standardize}
\title{Age-standardize samples}
\usage{
age_standardize(sample, std_pop, margin)
}
\arguments{
\item{sample}{Sample array to age-standardize}

\item{std_pop}{A vector of standard populations}

\item{margin}{The margin on which the age data is stratified}
}
\description{
Age-standardize samples
}
