% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{bind_samples}
\alias{bind_samples}
\title{Bind and name standardized/aggregated samples}
\usage{
bind_samples(sample, agg_sample, margin, new_name = NULL)
}
\arguments{
\item{sample}{Original sample array}

\item{agg_sample}{Newly aggregated/standardized array}

\item{margin}{The margin on which the groups of interest are stratified}

\item{new_name}{The name of the newly aggregated group}
}
\description{
Bind and name standardized/aggregated samples
}
