% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{group_aggregate}
\alias{group_aggregate}
\title{Aggregate samples}
\usage{
group_aggregate(sample, pop, margin)
}
\arguments{
\item{sample}{Sample array to age-standardize}

\item{pop}{A population array to be used for weighted averages}

\item{margin}{The margin on which the groups of interest are stratified}
}
\description{
Aggregate samples
}
