% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize CAR model}
\usage{
initialize_model(
  name,
  dir = getwd(),
  data,
  adjacency,
  inits = NULL,
  priors = NULL,
  model = c("mstcar", "ucar", "ustcar", "mcar"),
  method = c("binom", "pois"),
  m0 = 3,
  A = NULL,
  rho_up = FALSE,
  impute_lb = 1,
  impute_ub = 9,
  seed = 1234,
  .ignore_checks = FALSE
)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{dir}{Directory where model will live}

\item{data}{Dataset including mortality (Y) and population (n) information}

\item{adjacency}{Dataset including adjacency information}

\item{inits}{Optional list of initial conditions for each parameter}

\item{priors}{Optional list of priors for updates}

\item{model}{Run model as an MSTCAR/UCAR/USTCAR/MCAR model}

\item{method}{Run model with either Binomial data or Poisson data}

\item{m0}{Baseline neighbor count by region}

\item{A}{Describes intensity of smoothing between regions}

\item{rho_up}{Controls whether rho update is performed for USTCAR or MSTCAR models}

\item{impute_lb}{If counts between lb and ub are suppressed for privacy reasons, impute_lb is lower bound}

\item{impute_ub}{If counts between lb and ub are suppressed for privacy reasons, impute_ub is upper bound}

\item{seed}{Set of random seeds to use for data replication}

\item{.ignore_checks}{If set to TRUE, ignores data checks. Only use if you are certain that your input data is
correct and you are encountering bugs during setup}
}
\description{
Initialize CAR model
}
