% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{load_pop}
\alias{load_pop}
\title{Load population array}
\usage{
load_pop(name, dir)
}
\arguments{
\item{name}{Name of the model}

\item{dir}{Directory of the model}
}
\description{
Load population array
}
