% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{load_samples}
\alias{load_samples}
\title{Load MCMC samples}
\usage{
load_samples(name, dir, param = "theta", burn = 2000)
}
\arguments{
\item{name}{Name of model}

\item{dir}{Directory where model lives}

\item{param}{Which parameter samples to load}

\item{burn}{Numer of burn-in samples to discard}

\item{time}{TIme period of data to pull from. `last` gathers final time period data and `all` gathers all years of data.}
}
\description{
Load MCMC samples
}
