% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maexample}
\alias{maexample}
\title{Massachusetts Heart Attack Mortality Data}
\format{
`maexample`
a data frame with 117 rows and 10 variables:

\describe{
  \item{Notes}{Dataset notes, starts at line 85}
  \item{Year}{Year label}
  \item{Year.Code}{Year label}
  \item{County}{County name}
  \item{County.Code}{County FIPS code}
  \item{Gender}{Sex group}
  \item{Gender.Code}{Abbreviated sex group}
  \item{Deaths}{Death count}
  \item{Population}{Population count}
  \item{Crude.Rate}{Crude rate generated by death and population count}
}
}
\source{
<https://wonder.cdc.gov/cmf-icd9.html>
}
\usage{
maexample
}
\description{
An example dataset from CDC WONDER containing counts for Myocardial Infarction (ICD-9 code 410.0)
deaths in all 14 Massachusetts counties for individuals in 2 gender groups across 3 years.
}
\keyword{datasets}
