% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mamap}
\alias{mamap}
\title{Massachusetts Shapefile}
\format{
`mamap`
a data frame with 14 rows and 13 variables:

\describe{
  \item{STATEFP}{State FIPS code}
  \item{COUNTYFP}{County FIPS code}
  \item{COUNTYNS}{County GNIS code}
  \item{AFFGEOID}{Census Unique Identifier}
  \item{GEOID}{Census Unique Identifier, truncated}
  \item{NAME}{County Name}
  \item{NAMELSAD}{County LSAD code}
  \item{STUSPS}{Abbreviated State Name}
  \item{STATE_NAME}{State Name}
  \item{LSAD}{State LSAD code}
  \item{ALAND}{amount of land in square meters}
  \item{AWATER}{amount of water in square meters}
  \item{geometry}{shape data for each county}
}
}
\source{
<https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html>
}
\usage{
mamap
}
\description{
A dataset containing U.S. Census TIGER shape data for Massachusetts
}
\keyword{datasets}
