% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miadj}
\alias{miadj}
\title{Michigan Adjacency Data}
\format{
`miadj`
a list containing neighbor adjacency vectors.
}
\usage{
miadj
}
\description{
A dataset containing the adjacency structure for each county in Michigan
}
\keyword{datasets}
