% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miheart}
\alias{miheart}
\title{Michigan Heart Attack Mortality Data}
\format{
`miheart`
a list with two array objects:

\describe{
  \item{Y}{death count (0-1005)}
  \item{n}{population count (26-292828)}
}
}
\source{
<https://wonder.cdc.gov/cmf-icd9.html>
}
\usage{
miheart
}
\description{
A dataset containing counts for Myocardial Infarction (ICD-9 code 410.0)
deaths in all 83 Michigan counties for individuals in 6 age groups across 10 years. This
dataset also contains the corresponding population counts.
}
\keyword{datasets}
