% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_standardize.R
\name{pop_aggregate}
\alias{pop_aggregate}
\title{Aggregate population arrays}
\usage{
pop_aggregate(pop, margin)
}
\arguments{
\item{pop}{Population array}

\item{margin}{The margin on which the groups of interest are stratified}
}
\description{
Aggregate population arrays
}
