\name{rst-package}
\alias{rst-package}
\alias{rst}
\docType{package}
\title{
A new procedure to optimize the selection of groups in a classification tree
}
\description{
Agglomerative cluster analyses encompass many techniques, which have been widely used in various fields of science. In biology, and specifically ecology, datasets are generally highly variable and may contain outliers, which increase the difficulty to identify the number of clusters. This package implements a new criterion to determine statistically the optimal level of partition in a classification tree. The criterion robustness is tested against perturbated data (outliers) using an observation or variable containing randomly generated values.
}
\details{
\tabular{ll}{
Package: \tab rst\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-11\cr
License: \tab GPL v3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Sakina-Dorothée Ayata, Vincent Calgano, Lionel Guidi, Jean-Olivier Irisson

Maintained by Jean-Olivier Irisson <irisson@normalesup.org>
}
\references{
Guidi, L., Ibanez, F, Calcagno, V, and Beaugrand, G. A new procedure to optimize the selection of groups in a classification tree: applications for ecological data. Ecological Modelling, 220(4):451 - 461, 2009.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\keyword{ multivariate }
\keyword{ cluster }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
