% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sampler.R
\name{run_sampler}
\alias{run_sampler}
\title{Run Gibbs sampler}
\usage{
run_sampler(name, dir, .show_plots = TRUE)
}
\arguments{
\item{name}{Name of model and corresponding folder}

\item{dir}{Directory where model lives}
}
\description{
Run Gibbs sampler
}
