\name{RSTr-package}
\alias{RSTr-package}
\alias{RSTr}
\docType{package}
\title{
\packageTitle{RSTr}
}
\description{
\packageDescription{RSTr}
}
\details{
The RSTr package is designed to spatially smooth discrete small-area event rates using information from neighboring spatial regions. See `browseVignettes("RSTr")` for a series of tutorials on basic usage of the RSTr functions.
}
\author{
\packageAuthor{RSTr}

Maintainer: \packageMaintainer{RSTr}
}
\references{
Multivariate spatiotemporal modeling of age-specific stroke mortality. (2017). Ann. Appl. Stat., 11(4), 2165–2177. https://doi.org/10.1214/17-AOAS1068
}
\keyword{package}
