\name{RST-package}
\alias{RST-package}
\alias{RST}
\docType{package}
\title{
\packageTitle{RST}
}
\description{
\packageDescription{RST}
}
\details{
The RST package is designed to spatially smooth discrete small-area event rates using information from neighboring spatial regions. See `browseVignettes("RST")` for a series of tutorials on basic usage of the RST functions.
}
\author{
\packageAuthor{RST}

Maintainer: \packageMaintainer{RST}
}
\references{
Multivariate spatiotemporal modeling of age-specific stroke mortality. (2017). Ann. Appl. Stat., 11(4), 2165–2177. https://doi.org/10.1214/17-AOAS1068
}
\keyword{package}
